+++ paramiko/dsskey.py
@@ -25,7 +25,8 @@ from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import hashes, serialization
 from cryptography.hazmat.primitives.asymmetric import dsa
 from cryptography.hazmat.primitives.asymmetric.utils import (
-    decode_dss_signature, encode_dss_signature
+    decode_dss_signature,
+    encode_dss_signature,
 )
 
 from paramiko import util
@@ -42,8 +43,15 @@ class DSSKey(PKey):
     data.
     """
 
-    def __init__(self, msg=None, data=None, filename=None, password=None,
-                 vals=None, file_obj=None):
+    def __init__(
+        self,
+        msg=None,
+        data=None,
+        filename=None,
+        password=None,
+        vals=None,
+        file_obj=None,
+    ):
         self.p = None
         self.q = None
         self.g = None
@@ -63,8 +71,8 @@ class DSSKey(PKey):
         else:
             self._check_type_and_load_cert(
                 msg=msg,
-                key_type='ssh-dss',
-                cert_type='ssh-dss-cert-v01@openssh.com',
+                key_type="ssh-dss",
+                cert_type="ssh-dss-cert-v01@openssh.com",
             )
             self.p = msg.get_mpint()
             self.q = msg.get_mpint()
@@ -74,7 +82,7 @@ class DSSKey(PKey):
 
     def asbytes(self):
         m = Message()
-        m.add_string('ssh-dss')
+        m.add_string("ssh-dss")
         m.add_mpint(self.p)
         m.add_mpint(self.q)
         m.add_mpint(self.g)
@@ -88,7 +96,7 @@ class DSSKey(PKey):
         return hash((self.get_name(), self.p, self.q, self.g, self.y))
 
     def get_name(self):
-        return 'ssh-dss'
+        return "ssh-dss"
 
     def get_bits(self):
         return self.size
@@ -102,17 +110,15 @@ class DSSKey(PKey):
             public_numbers=dsa.DSAPublicNumbers(
                 y=self.y,
                 parameter_numbers=dsa.DSAParameterNumbers(
-                    p=self.p,
-                    q=self.q,
-                    g=self.g
-                )
-            )
+                    p=self.p, q=self.q, g=self.g
+                ),
+            ),
         ).private_key(backend=default_backend())
         sig = key.sign(data, hashes.SHA1())
         r, s = decode_dss_signature(sig)
 
         m = Message()
-        m.add_string('ssh-dss')
+        m.add_string("ssh-dss")
         # apparently, in rare cases, r or s may be shorter than 20 bytes!
         rstr = util.deflate_long(r, 0)
         sstr = util.deflate_long(s, 0)
@@ -129,7 +135,7 @@ class DSSKey(PKey):
             sig = msg.asbytes()
         else:
             kind = msg.get_text()
-            if kind != 'ssh-dss':
+            if kind != "ssh-dss":
                 return 0
             sig = msg.get_binary()
 
@@ -142,10 +148,8 @@ class DSSKey(PKey):
         key = dsa.DSAPublicNumbers(
             y=self.y,
             parameter_numbers=dsa.DSAParameterNumbers(
-                p=self.p,
-                q=self.q,
-                g=self.g
-            )
+                p=self.p, q=self.q, g=self.g
+            ),
         ).public_key(backend=default_backend())
         try:
             key.verify(signature, data, hashes.SHA1())
@@ -160,18 +164,16 @@ class DSSKey(PKey):
             public_numbers=dsa.DSAPublicNumbers(
                 y=self.y,
                 parameter_numbers=dsa.DSAParameterNumbers(
-                    p=self.p,
-                    q=self.q,
-                    g=self.g
-                )
-            )
+                    p=self.p, q=self.q, g=self.g
+                ),
+            ),
         ).private_key(backend=default_backend())
 
         self._write_private_key_file(
             filename,
             key,
             serialization.PrivateFormat.TraditionalOpenSSL,
-            password=password
+            password=password,
         )
 
     def write_private_key(self, file_obj, password=None):
@@ -180,18 +182,16 @@ class DSSKey(PKey):
             public_numbers=dsa.DSAPublicNumbers(
                 y=self.y,
                 parameter_numbers=dsa.DSAParameterNumbers(
-                    p=self.p,
-                    q=self.q,
-                    g=self.g
-                )
-            )
+                    p=self.p, q=self.q, g=self.g
+                ),
+            ),
         ).private_key(backend=default_backend())
 
         self._write_private_key(
             file_obj,
             key,
             serialization.PrivateFormat.TraditionalOpenSSL,
-            password=password
+            password=password,
         )
 
     @staticmethod
@@ -207,23 +207,25 @@ class DSSKey(PKey):
         numbers = dsa.generate_private_key(
             bits, backend=default_backend()
         ).private_numbers()
-        key = DSSKey(vals=(
-            numbers.public_numbers.parameter_numbers.p,
-            numbers.public_numbers.parameter_numbers.q,
-            numbers.public_numbers.parameter_numbers.g,
-            numbers.public_numbers.y
-        ))
+        key = DSSKey(
+            vals=(
+                numbers.public_numbers.parameter_numbers.p,
+                numbers.public_numbers.parameter_numbers.q,
+                numbers.public_numbers.parameter_numbers.g,
+                numbers.public_numbers.y,
+            )
+        )
         key.x = numbers.x
         return key
 
     # ...internals...
 
     def _from_private_key_file(self, filename, password):
-        data = self._read_private_key_file('DSA', filename, password)
+        data = self._read_private_key_file("DSA", filename, password)
         self._decode_key(data)
 
     def _from_private_key(self, file_obj, password):
-        data = self._read_private_key('DSA', file_obj, password)
+        data = self._read_private_key("DSA", file_obj, password)
         self._decode_key(data)
 
     def _decode_key(self, data):
@@ -232,14 +234,11 @@ class DSSKey(PKey):
         try:
             keylist = BER(data).decode()
         except BERException as e:
-            raise SSHException('Unable to parse key file: ' + str(e))
-        if (
-            type(keylist) is not list or
-            len(keylist) < 6 or
-            keylist[0] != 0
-        ):
+            raise SSHException("Unable to parse key file: " + str(e))
+        if type(keylist) is not list or len(keylist) < 6 or keylist[0] != 0:
             raise SSHException(
-                'not a valid DSA private key file (bad ber encoding)')
+                "not a valid DSA private key file (bad ber encoding)"
+            )
         self.p = keylist[1]
         self.q = keylist[2]
         self.g = keylist[3]